requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
defusedxml==0.7.1
Deprecated==1.3.1
dnspython==2.8.0
eventlet==0.40.4
exceptiongroup==1.3.1
fastapi==0.122.0
filelock==3.20.0
Flask==3.1.2
fsspec==2025.10.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
Levenshtein==0.27.3
limits==5.6.0
MarkupSafe==3.0.3
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1747544640217/work/dist/numpy-2.2.6-cp310-cp310-linux_x86_64.whl#sha256=d6d964caeef85d00073d27cd62b46883d275b3d8162f723f0fcabbd0b3cc3f9d
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
opencv-python==4.10.0
opencv-python-headless==4.10.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pyclipper==1.3.0.post6
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==4.30.0
python-doctr==1.0.0
python-Levenshtein==0.27.3
python-multipart==0.0.20
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
scipy==1.15.3
shapely==2.1.2
six==1.17.0
slowapi==0.1.9
starlette==0.50.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
validators==0.35.0
Werkzeug==3.1.4
wrapt==2.0.1
